GH-63 : Moved entities and repositories and common things in common library project
