+++ pom.xml
@@ -18,10 +18,9 @@
 		<properties-maven-plugin.version>1.2.1</properties-maven-plugin.version>
 		<sonar-maven-plugin.version>3.11.0.3922</sonar-maven-plugin.version>
 		<jacoco-maven-plugin.version>0.8.12</jacoco-maven-plugin.version>
-		<ulid-creator.version>5.2.3</ulid-creator.version>
-		<jsoup.version>1.17.2</jsoup.version>
 		<greenmail-junit5.version>2.0.1</greenmail-junit5.version>
 		<firebase-admin.version>9.2.0</firebase-admin.version>
+		<ihmsservicecommon.version>0.0.1</ihmsservicecommon.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -36,6 +35,12 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-validation</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>com.amalvadkar</groupId>
+			<artifactId>ihmsservicecommon</artifactId>
+			<version>${ihmsservicecommon.version}</version>
 		</dependency>
 
 		<dependency>
@@ -90,11 +95,7 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
 		</dependency>
-		<dependency>
-			<groupId>org.jsoup</groupId>
-			<artifactId>jsoup</artifactId>
-			<version>${jsoup.version}</version>
-		</dependency>
+
 		<dependency>
 			<groupId>com.icegreen</groupId>
 			<artifactId>greenmail-junit5</artifactId>
