--- src/main/java/com/amalvadkar/ihms/common/entities/PolicyCategoryEntity.java
@@ -1,24 +1,0 @@
-package com.amalvadkar.ihms.common.entities;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-
-
-@Entity
-@Table(name = "policy_category")
-@Getter
-@Setter
-public class PolicyCategoryEntity extends AbstractIdBaseEntity {
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "display_order", nullable = false)
-    private BigDecimal displayOrder;
-
-}
