--- src/main/java/com/amalvadkar/ihms/common/entities/UserEntity.java
@@ -1,25 +1,0 @@
-package com.amalvadkar.ihms.common.entities;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "users")
-@Getter
-@Setter
-public class UserEntity extends AbstractIdBaseEntity {
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "email", nullable = false)
-    private String email;
-
-    @Column(name = "active", nullable = false)
-    private Boolean active;
-
-
-}
