--- src/main/java/com/amalvadkar/ihms/common/enums/CategoryEnum.java
@@ -1,7 +1,0 @@
-package com.amalvadkar.ihms.common.enums;
-
-public enum CategoryEnum {
-
-    FEEDBACK
-
-}
