--- src/main/java/com/amalvadkar/ihms/common/enums/CountryIsoCodeEnum.java
@@ -1,6 +1,0 @@
-package com.amalvadkar.ihms.common.enums;
-
-public enum CountryIsoCodeEnum {
-    IND,
-    USA
-}
