--- src/main/java/com/amalvadkar/ihms/common/enums/OptionalStatusEnum.java
@@ -1,7 +1,0 @@
-package com.amalvadkar.ihms.common.enums;
-
-public enum OptionalStatusEnum {
-    Yes,
-
-    No
-}
