--- src/main/java/com/amalvadkar/ihms/common/models/request/FetchFilesRequest.java
@@ -1,8 +1,0 @@
-package com.amalvadkar.ihms.common.models.request;
-
-import jakarta.validation.constraints.NotEmpty;
-
-public record FetchFilesRequest(
-        @NotEmpty(message = "recordId is required")
-        String recordId
-) {}
