--- src/main/java/com/amalvadkar/ihms/common/models/request/ViewFileRequest.java
@@ -1,10 +1,0 @@
-package com.amalvadkar.ihms.common.models.request;
-
-import jakarta.validation.constraints.NotEmpty;
-
-public record ViewFileRequest(
-
-        @NotEmpty(message = "fileId is required")
-        String fileId
-) {
-}
