--- src/main/java/com/amalvadkar/ihms/common/models/response/FetchFilesContentResModel.java
@@ -1,6 +1,0 @@
-package com.amalvadkar.ihms.common.models.response;
-
-import java.time.Instant;
-
-public record FetchFilesContentResModel(String fileId, String fileName, String uploadedBy, Instant uploadedOn) {
-}
