--- src/main/java/com/amalvadkar/ihms/common/models/response/FetchFilesResModel.java
@@ -1,6 +1,0 @@
-package com.amalvadkar.ihms.common.models.response;
-
-import java.util.List;
-
-public record FetchFilesResModel(List<FetchFilesContentResModel> content) {
-}
