--- src/main/java/com/amalvadkar/ihms/common/models/response/KeyValueResModel.java
@@ -1,4 +1,0 @@
-package com.amalvadkar.ihms.common.models.response;
-
-public record KeyValueResModel(Long key, String value) {
-}
