--- src/main/java/com/amalvadkar/ihms/common/repositories/FeedBackRepository.java
@@ -1,7 +1,0 @@
-package com.amalvadkar.ihms.common.repositories;
-
-import com.amalvadkar.ihms.common.entities.FeedBackEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FeedBackRepository extends JpaRepository<FeedBackEntity, String> {
-}
