--- src/main/java/com/amalvadkar/ihms/feedback/enums/FeedbackStatusEnum.java
@@ -1,10 +1,0 @@
-package com.amalvadkar.ihms.feedback.enums;
-
-public enum FeedbackStatusEnum {
-
-    REJECTED,
-
-    ACCEPTED,
-
-    SUBMITTED
-}
