+++ src/main/java/com/amalvadkar/ihms/files/models/response/FetchFilesResModel.java
@@ -1,0 +1,8 @@
+package com.amalvadkar.ihms.files.models.response;
+
+import com.amalvadkar.ihms.common.models.dto.FetchFilesContentDTO;
+
+import java.util.List;
+
+public record FetchFilesResModel(List<FetchFilesContentDTO> content) {
+}
