+++ src/main/java/com/amalvadkar/ihms/files/request/FetchFilesRequest.java
@@ -1,0 +1,8 @@
+package com.amalvadkar.ihms.files.request;
+
+import jakarta.validation.constraints.NotEmpty;
+
+public record FetchFilesRequest(
+        @NotEmpty(message = "recordId is required")
+        String recordId
+) {}
