+++ src/main/java/com/amalvadkar/ihms/files/request/ViewFileRequest.java
@@ -1,0 +1,10 @@
+package com.amalvadkar.ihms.files.request;
+
+import jakarta.validation.constraints.NotEmpty;
+
+public record ViewFileRequest(
+
+        @NotEmpty(message = "fileId is required")
+        String fileId
+) {
+}
