+++ src/test/resources/add_test_user_and_delete_all_users_except_system_user.sql
@@ -1,7 +1,7 @@
 SET FOREIGN_KEY_CHECKS = 0;
 
 delete from users where id > 1;
-insert into users(id,name, email, created_on)
-values (2, "Test" ,"it@test.com", utc_timestamp());
+insert into users(id,name, email,role_id, created_on)
+values (2, "Test" ,"it@test.com","01HXXBNZSEKH1RD14TTVDCZG4K", utc_timestamp());
 
 SET FOREIGN_KEY_CHECKS = 1;
